+++ petuhov.kirill/S1/readinput.hpp
@@ -6,7 +6,7 @@
 
 namespace petuhov
 {
-  void readInput(List< std::pair< std::string, List< unsigned long long > > >& sequences, std::istream& input);
+  void readInput(List< std::pair< std::string, List< size_t > > >& sequences, std::istream& input);
 }
 
 #endif
